+++ plugins/dom.js
@@ -1,0 +1,15 @@
+import { DOMParser } from "../deps/dom.js";
+
+export default function (callback) {
+  const parser = new DOMParser();
+
+  return (site) => {
+    site.afterRender([".html"], transform);
+
+    async function transform(page) {
+      const document = parser.parseFromString(page.rendered, "text/html");
+      callback(document, page);
+      page.rendered = document.documentElement.outerHTML;
+    }
+  };
+}