+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 - Rebuild and reload on change `_data` files
+- Explorer returns wrong results
 
 ## [0.2.3] - 2020-09-14
 ### Fixed