+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- New plugin `svg` to optimize svg files
+
 ### Fixed
 - Refactored source load and reload
 - Explorer returns wrong results